fix: constrain total_score_pct to NUMERIC(4,1) to stop storing raw float noise
The column was an unconstrained NUMERIC, so Postgres stored the exact
binary expansion of the Python float app code already rounds to one
decimal place (e.g. 88.4 became 88.400000000000005684...) instead of
the rounded value. NUMERIC(4,1) matches every round(x, 1) call site
that computes this value (excel_handler.py, reviews.py) and makes
Postgres itself enforce the scale on every write from now on.
Migration re-rounds existing rows via ALTER COLUMN ... USING round(),
verified end-to-end against a throwaway Postgres instance (bug
reproduced, migration applied, value corrected to 88.4; downgrade
also verified).

+++ backend/app/db/models.py
@@ -29,7 +29,11 @@ class PlatformReview(Base):
     project_name: Mapped[str] = mapped_column(String, nullable=False)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True))
     completed_at: Mapped[Optional[datetime]] = mapped_column(DateTime(timezone=True), nullable=True)
-    total_score_pct: Mapped[Optional[float]] = mapped_column(Numeric, nullable=True)
+    # Explicit precision/scale (0.0-100.0, matching every round(x, 1) call
+    # site that computes this value) -- an unconstrained Numeric stores the
+    # exact binary expansion of the Python float (e.g. 88.4 becomes
+    # 88.400000000000005684...), since Postgres has no scale to round to.
+    total_score_pct: Mapped[Optional[float]] = mapped_column(Numeric(4, 1), nullable=True)
     llm_provider: Mapped[str] = mapped_column(String, nullable=False)
     llm_model: Mapped[Optional[str]] = mapped_column(String, nullable=True)
     compile_check_mode: Mapped[str] = mapped_column(String, nullable=False)
